requirements.txt
Django==1.4.2
Pillow==1.7.8
boto==2.9.2
dj-database-url==0.2.1
django-facebook==4.2.11
django-guardian==1.1.0.beta
django-mongonaut==0.2.17
django-s3-folder-storage==0.1
django-storages==1.1.8
django-userena==1.2.0
easy-thumbnails==1.2
gunicorn==0.17.2
ipdb==0.7
ipython==0.13.1
mongoengine==0.7.9
psycopg2==2.4.6
pymongo==2.4.2
readline==6.2.4.1
simplejson==2.6.2
wsgiref==0.1.2
git+git://github.com/feliperyan/django-mongodbforms.git
